+++ src/skin_color_classifier.py
@@ -1,5 +1,5 @@
-import RFClassifier as rf
-import KNNClassifier as knn
+import rf_classifier as rf
+import knn_classifier as knn
 import skin_detection as sd
 import numpy as np
 import cv2
@@ -10,28 +10,32 @@ class SkinColorClassifier():
     # min_img, max_img is non masked image
     # both image ensures that hand image exists 
     # and location of the hand image is different
-    def __init__(self, min_img, max_img):
+    def __init__(self, img1, img2, classifier="knn"):
 
         # mask both images and classify 
-        pos, neg = self.collect_data(min_img, max_img)
-        #self.classifier = rf.RFClassifier(pos, neg)
-        self.classifier = knn.KNNClassifier(pos, neg)
-     
-    def collect_data(self, min_img, max_img):
-        
-        min_mask = sd.mask_skin(min_img)
-        max_mask = sd.mask_skin(max_img)
+        pos, neg = self.collect_data(img1, img2)
+        if classifier == "knn":
+            self.classifier = knn.KNNClassifier(pos, neg)
+        elif classifier == "rf":
+            self.classifier = rf.RFClassifier(pos, neg)
+        else:
+            print("classifier argument not allowed")
 
-        and_mask = cv2.bitwise_and(min_mask, max_mask)
+    def collect_data(self, img1, img2):
+        
+        mask1 = sd.mask_skin(img1)
+        mask2 = sd.mask_skin(img2)
+
+        and_mask = cv2.bitwise_and(mask1, mask2)
         
         # mask neg image of min_image
-        neg = self.collect_neg_data(min_img, and_mask)
+        neg = self.collect_neg_data(img1, and_mask)
 
         # mask pos image of min_image
-        pos = self.collect_pos_data(min_img, min_mask, and_mask)
+        pos = self.collect_pos_data(img1, mask1, and_mask)
         # mask pos image of max_image
-        #pos = np.append(pos, \
-        #    self.collect_pos_data(max_img, max_mask, and_mask), axis=0)
+        pos = np.append(pos, \
+            self.collect_pos_data(img2, mask2, and_mask), axis=0)
 
         return pos, neg
 
@@ -54,7 +58,7 @@ class SkinColorClassifier():
         pos_bool = np.nonzero(pos_mask)
         return img[pos_bool[0]]    
       
-    def mask_with_classifier(self, img):
+    def mask_image(self, img):
 
         #blur_k_size = (11, 11)        
         #img = cv2.blur(img, blur_k_size)
@@ -80,8 +84,7 @@ class SkinColorClassifier():
 
 if __name__ == '__main__':
 
-
-    # test data
+    # test data 
     dir1 = "../test-data/img_min.png"
     dir2 = "../test-data/img_max.png"
     test_dir = "../test-data/img_test.png"
@@ -96,7 +99,7 @@ if __name__ == '__main__':
    
     start_time = time.time()
     scc = SkinColorClassifier(min_img, max_img)
-    out = scc.mask_with_classifier(test_img)
+    out = scc.mask_image(test_img)
     print("Time duration: " + str(time.time() - start_time))
     
     cv2.imshow('fixed mask', sd.mask_skin(copy))
